# apps/go/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Runa Backend API
  version: 0.6.0
  description: |
    HTTP API for the Runa (moon-themed) app. Served under the /api/v1 base path.
    Clients are injected host+port only and append /api/v1/... themselves.

    Authentication: the /api/v1/auth/* and /healthz routes are public; GET /me and
    all /api/v1/diary*, /api/v1/today, /api/v1/songs* routes require a Bearer
    access token. Access tokens are short-lived HS256 JWTs; refresh tokens are
    long-lived opaque strings, stored server-side only as a hash and rotated on
    every /auth/refresh. The /api/v1/admin/* seed endpoints use a separate shared
    admin token (X-Admin-Token), not a user session.

    Diary sync: entries are offline-first. The client generates client_id (a UUID)
    so an offline create can be POSTed idempotently on reconnect without
    duplicating. GET /diary/sync?since= returns changes (including deletion
    tombstones) after a watermark; conflicts resolve last-write-wins by updated_at.

    Today: GET /today returns the day's curated quote and song (exact date match;
    either may be null). The moon phase — the home's third element — is NOT served
    here; it is computed on the client in shared code (Meeus-simplified) so the
    home works fully offline. GET /songs is the newest-first song archive and
    POST /songs/{id}/played records play history.

    Gallery (presigned-URL scheme): image BYTES never traverse this API. To add an
...
servers:
...
paths:
...
  /api/v1/me:
    get:
      operationId: getMe
      summary: Get the authenticated user (protected)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The authenticated user's record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing, invalid or expired access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMe
      summary: Update the authenticated user's display name (protected)
      description: |
        Replaces display_name. The name is trimmed; an empty result is rejected
        (400) and it may be at most 50 characters.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: The updated user record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMe
      summary: Permanently delete the authenticated account (protected)
      description: |
        Hard-deletes the user and every row that cascades from it — refresh
        tokens, diary entries, gallery image rows and song history — in a single
        statement. The user's stored image objects are purged from object storage
        asynchronously (best-effort). This is irreversible: there is no recovery
        window. Refresh tokens are revoked by the cascade; a still-valid access
        token stops working on its next request (the user no longer exists), so the
        residual window is bounded by the short access-token lifetime.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token returned by any auth endpoint.
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: |
        Shared admin token for the curated seed endpoints. Set server-side via
        ADMIN_API_TOKEN; when unset the admin endpoints are disabled (403).
  responses:
    ValidationError:
      description: The request body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: validation_error
              message: validation failed
              details:
                - field: email
                  message: must be a valid email address
...
    Unauthorized:
      description: Missing, invalid or expired access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: authentication required
...
  schemas:
...
    User:
      type: object
      required:
        - id
        - display_name
        - auth_provider
        - is_premium
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        display_name:
          type: string
        auth_provider:
          type: string
          enum: [email, apple, google]
        is_premium:
          type: boolean
        premium_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    UpdateMeRequest:
      type: object
      required:
        - display_name
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 50
          description: New display name; trimmed server-side, must be non-empty.
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - validation_error
                - invalid_credentials
                - email_taken
                - unauthorized
                - token_expired
                - token_invalid
                - provider_verification_failed
                - rate_limited
                - not_found
                - internal_error
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...